webgui: introduce BaseHandler class
It is base class, derived from CefClient, which mainly should handle 
messaging between JS client and web canvas. Such communication required
both in GUI and off-screen mode
