+++ gui/canvaspainter/CMakeLists.txt
@@ -14,7 +14,7 @@ if(DEFINED ENV{CEF_PATH})
 if(EXISTS $ENV{CEF_PATH})
    include_directories($ENV{CEF_PATH})
 #   add_definitions(-DWEBGUI_WITH_CEF)
-   ROOT_GLOB_SOURCES(sources v7/src/*.cxx v7/cef/simple_app.cxx v7/cef/osr_handler.cxx v7/cef/simple_handler.cxx v7/cef/simple_handler_linux.cxx)
+   ROOT_GLOB_SOURCES(sources v7/src/*.cxx v7/cef/simple_app.cxx v7/cef/base_handler.cxx v7/cef/osr_handler.cxx v7/cef/simple_handler.cxx v7/cef/simple_handler_linux.cxx)
    configure_file($ENV{CEF_PATH}/Resources/icudtl.dat "../../bin/icudtl.dat" COPYONLY)
    configure_file($ENV{CEF_PATH}/Release/natives_blob.bin "../../bin/natives_blob.bin" COPYONLY)
    configure_file($ENV{CEF_PATH}/Release/snapshot_blob.bin "../../bin/snapshot_blob.bin" COPYONLY)
